# txhelpers: store previously-computed ultimate subsidy values