--- a/txhelpers/subsidy.go
+++ b/txhelpers/subsidy.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2015-2018 The Decred developers
+// Copyright (c) 2015-2019 The Decred developers
 // Copyright (c) 2013-2015 The btcsuite developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
@@ -10,12 +10,21 @@ import (
 	"github.com/decred/dcrd/chaincfg"
 )
 
+// ultimateSubsidies stores ultimate subsidy values computed by UltimateSubsidy.
+var ultimateSubsidies = map[*chaincfg.Params]int64{}
+
 // UltimateSubsidy computes the total subsidy over the entire subsidy
 // distribution period of the network.
 func UltimateSubsidy(params *chaincfg.Params) int64 {
+	// Check previously computed ultimate subsidies.
+	totalSubsidy, ok := ultimateSubsidies[params]
+	if ok {
+		return totalSubsidy
+	}
+
 	subsidyCache := blockchain.NewSubsidyCache(0, params)
 
-	totalSubsidy := params.BlockOneSubsidy()
+	totalSubsidy = params.BlockOneSubsidy()
 	for i := int64(0); ; i++ {
 		// Genesis block or first block.
 		if i <= 1 {
@@ -49,6 +58,10 @@ func UltimateSubsidy(params *chaincfg.Params) int64 {
 			}
 		}
 	}
+
+	// Update the ultimate subsidy store.
+	ultimateSubsidies[params] = totalSubsidy
+
 	return totalSubsidy
 }