The "rotation" API attack shows that right below SQLite queries,
`txhelpers.UltimateSubsidy` dominates CPU use:

```
Profiling CPU now, will take 30 secs...
Profile dump saved to: /tmp/profile382211380
Profiling dump saved to: /tmp/profile382211380
Binary file saved to: /tmp/binary999078659
File: binary999078659
Build ID: dac764d5609000fe955fb95b2afecec454b17e47
Type: cpu
Time: Mar 15, 2019 at 8:33am (CDT)
Duration: 30.14s, Total samples = 1.38mins (275.41%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 75.69s, 91.19% of 83s total
Dropped 543 nodes (cum <= 0.41s)
Showing top 10 nodes out of 126
      flat  flat%   sum%        cum   cum%
    48.31s 58.20% 58.20%     48.32s 58.22%  runtime.cgocall
    23.13s 27.87% 86.07%     23.32s 28.10%  github.com/decred/dcrdata/v4/txhelpers.UltimateSubsidy
     0.89s  1.07% 87.14%      0.91s  1.10%  syscall.Syscall
     0.76s  0.92% 88.06%      1.38s  1.66%  runtime.mapassign
     0.56s  0.67% 88.73%      0.56s  0.67%  runtime.aeshashbody
...
```

These changes add a package-level store for previously computed
ultimate subsidy values.  `txhelpers.UltimateSubsidy` uses it to avoid
recomputing the same value repeatedly.

Before:

```
BenchmarkUltimateSubsidy-32    	    20	  85572558 ns/op	 190943 B/op	   111 allocs/op
```

After:

```
BenchmarkUltimateSubsidy-32    	300000000	  4.52 ns/op	    0 B/op	      0 allocs/op
```